Change default format to XML
The default format for ActiveResource objects was XML prior to 3.1,
but has since changed. Unfortunately, additional work would be needed
to correctly parse the JSON coming back from the API, as it includes
the root class name in the returned object. To simplify things, it's
best to continue using XML.